pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-auth>=2.47.0",
    "httpx>=0.28.1",
    "image>=1.5.33",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "cryptography>=41.0.0",
    "python-multipart>=0.0.6",
    "PyJWT>=2.0.0",
    "pydantic-settings>=2.12.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config --cov=app --cov-report=term-missing --cov-report=html --cov-report=xml"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "security: marks tests as security tests",
    "vertex: marks tests for Vertex AI models",
    "gemini: marks tests for Gemini models",
    "imagen: marks tests for Imagen models",
    "veo: marks tests for Veo models",
]

